InfluxdbCommonWriter: abstract InfluxdbWriter
refs #8711
