+++ etc/icinga2/features-available/influxdb.conf
@@ -1,6 +1,6 @@
 /**
  * The InfluxdbWriter type writes check result metrics and
- * performance data to an InfluxDB HTTP API
+ * performance data to an InfluxDB v1 HTTP API
  */
 
 object InfluxdbWriter "influxdb" {
