+++ lib/perfdata/influxdbwriter.hpp
@@ -4,19 +4,12 @@
 #define INFLUXDBWRITER_H
 
 #include "perfdata/influxdbwriter-ti.hpp"
-#include "icinga/service.hpp"
-#include "base/configobject.hpp"
-#include "base/tcpsocket.hpp"
-#include "base/timer.hpp"
-#include "base/tlsstream.hpp"
-#include "base/workqueue.hpp"
-#include <fstream>
 
 namespace icinga
 {
 
 /**
- * An Icinga InfluxDB writer.
+ * An Icinga InfluxDB v1 writer.
  *
  * @ingroup perfdata
  */
@@ -28,35 +21,9 @@ public:
 
 	static void StatsFunc(const Dictionary::Ptr& status, const Array::Ptr& perfdata);
 
-	void ValidateHostTemplate(const Lazy<Dictionary::Ptr>& lvalue, const ValidationUtils& utils) override;
-	void ValidateServiceTemplate(const Lazy<Dictionary::Ptr>& lvalue, const ValidationUtils& utils) override;
-
 protected:
-	void OnConfigLoaded() override;
-	void Resume() override;
-	void Pause() override;
-
-private:
-	WorkQueue m_WorkQueue{10000000, 1};
-	Timer::Ptr m_FlushTimer;
-	std::vector<String> m_DataBuffer;
-
-	void CheckResultHandler(const Checkable::Ptr& checkable, const CheckResult::Ptr& cr);
-	void CheckResultHandlerWQ(const Checkable::Ptr& checkable, const CheckResult::Ptr& cr);
-	void SendMetric(const Checkable::Ptr& checkable, const Dictionary::Ptr& tmpl,
-		const String& label, const Dictionary::Ptr& fields, double ts);
-	void FlushTimeout();
-	void FlushTimeoutWQ();
-	void Flush();
-
-	static String EscapeKeyOrTagValue(const String& str);
-	static String EscapeValue(const Value& value);
-
-	OptionalTlsStream Connect();
-
-	void AssertOnWorkQueue();
-
-	void ExceptionHandler(boost::exception_ptr exp);
+	boost::beast::http::request<boost::beast::http::string_body> AssembleRequest(String body) override;
+	Url::Ptr AssembleUrl() override;
 };
 
 }
